% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICD10to9.R
\name{ICD10to9}
\alias{ICD10to9}
\title{ICD Transformation and mortality calculation}
\usage{
ICD10to9(sample, dfr_col, Diag.ICD10, dictionary, pop.data, age.cut, YPLL.yr)
}
\arguments{
\item{sample}{A dataset of death records.\cr
It should contain 4 columns: 1. SEXO, 2. EDAD_TIPO, 3. EDAD_CANT, 4. dfr_col.\cr
SEXO: 1=males, 2=females; EDAD_TIPO: Type of age measure, 1=years, 2= months, 3 =days, 4=hours and minutes; EDAD_CANT: Quantity of age (in EDAD_TIPO measure); dfr_col: ICD code for death reason of this individual.}

\item{dfr_col}{The name of ICD9/10 code column in SAMPLE dataset.}

\item{dictionary}{A dataset with two columns which users prepared for ICD transformation.\cr
It should contain 2 columns: 1. ICD10, 2. ICD9.\cr
ICD10: ICD10 code, ICD9: corresponding codes in ICD9.\cr
IMPORTANCE: this package can only identify ICD9 codes included in this DICTIONARY dataset. Make sure that ICD9 code in the DICTIONARY has all possible formats of ICD9 in the SAMPLE dataset.\cr}

\item{age.cut}{User-defined age group cuts.}

\item{YPLL.yr}{Expected years of life.}

\item{Diag.ICD9}{Diagnosis ICD10 code(s) for disease(s) of interest.\cr
It should be the same format as 'ICD10' in DICTIONARY dataset.}

\item{population}{A Dataset of population.\cr
It should contain 3 columns: 1. Sex, 2. Age, 3. Value.\cr
Sex: 'Both Sexes', 'Male', 'Female'; Age: numeric; Value: numeric, the number of individuals of a certain age and gender}
}
\value{
I. unadjusted.death.rate$total.death.rate: unadjusted death rate for total population by user-defined age group;\cr
II. unadjusted.death.rate$male.death.rate: unadjusted death rate for males by user-defined age group;\cr
III. unadjusted.death.rate$female.death.rate: unadjusted death rate for females by user-defined age group;\cr
IV. direct: age-adjusted death rates using direct method by gender\cr
V. indirect: standardized mortality ratio (SMR) by gender\cr
VI. YPLL: Years of potential life lost
}
\description{
This function is able to identify ICD9 code(s) matched ICD10 code(s) of interest, and then
 calculate the disease-specific age-adjusted mortality rate and
years of potential life lost (YPLL).
}
\details{
Three datasets should be prepared for this function. \cr
     1. a death dataset which contains gender, age, and ICD code of death reason;\cr
     2. a population dataset which contains numbers of people for each gender and age; \cr
     3. a dictionary which contains ICD10 codes and their corresponding ICD9 codes.\cr \cr
After inputting any ICD10 code, 'ICD10to9' will search for death records of both ICD10 and corresponding ICD9 code.
And then, the calculation will be based on these death records. One limitation of the current version is
that the datasets should be preprocessed according to the description in Arguments section. Our team is trying to make it more flexible. The package will be updated later.\cr \cr
Within this package, a sample dataset which is death records in Chile, Chile population dataset from UN, and
a dictionary for cardiovascular disease are provided.
}
\examples{
data(Chile.death.1994.ICD9)
data(Chile.population.1998)
data(ICD10to9CVD)
ICD10to9(sample = Chile.death.1994.ICD9, dfr_col = 'DIAG1', Diag.ICD10 = c('I21.0'), 
dictionary = ICD10to9CVD, pop.data = Chile.population.1998, age.cut = c(1, 5, 15, 25, 45, 65, 80), YPLL.yr = 80)
}
